print passing and failing tests

+++ array.c
@@ -15,12 +15,14 @@ struct Array CreateArray(size_t element_size) {
 void PushArray(struct Array *array, void *element) {
   if (array->used == array->capacity) {
     array->capacity *= 2;
-    array->elements = realloc(array->elements, array->capacity * array->element_size);
+    array->elements =
+        realloc(array->elements, array->capacity * array->element_size);
   }
-  memcpy(array->elements + (array->used * array->element_size), element, array->element_size);
+  memcpy(array->elements + (array->used * array->element_size), element,
+         array->element_size);
   array->used++;
 }
 
-void *GetElementArray(struct Array *array, size_t index) {
+void *GetElementArray(struct Array *array, int index) {
   return (array->elements + (array->element_size * index));
 }
